Completed basic DocDb Persistence functionality.

+++ Xigadee.Core/Persistence/Server/DocumentDb/Helper/RestBase.cs
@@ -208,8 +208,8 @@ namespace Xigadee
 
                     if (rs.Response.Headers.Contains("x-ms-request-charge"))
                     {
-                        long charge;
-                        if (long.TryParse(rs.Response.Headers.GetValues("x-ms-request-charge").FirstOrDefault(), out charge))
+                        double charge;
+                        if (double.TryParse(rs.Response.Headers.GetValues("x-ms-request-charge").FirstOrDefault(), out charge))
                             rs.ResourceCharge = charge;
                     }
